pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yucode-agent"
version = "0.2.0"
description = "Python-first coding agent runtime and VS Code bridge with multi-worker orchestration"
requires-python = ">=3.10"
readme = "README.md"
keywords = ["coding-agent", "llm", "ai", "orchestration", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = []

[project.urls]
Homepage = "https://github.com/user/yucode"
Repository = "https://github.com/user/yucode"
Documentation = "https://github.com/user/yucode#readme"
Changelog = "https://github.com/user/yucode/blob/main/CHANGELOG.md"

[project.optional-dependencies]
server = ["aiohttp>=3.9"]
excel = ["openpyxl>=3.1"]
word = ["python-docx>=1.1"]
pptx = ["python-pptx>=1.0"]
pdf = ["pdfplumber>=0.10"]
finance = ["yfinance>=0.2", "pandas>=2.0"]
office = ["openpyxl>=3.1", "python-docx>=1.1", "python-pptx>=1.0", "pdfplumber>=0.10"]
all = [
    "aiohttp>=3.9",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "pdfplumber>=0.10",
    "yfinance>=0.2",
    "pandas>=2.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
yucode = "coding_agent.interface.cli:main"

[tool.setuptools.packages.find]
include = ["coding_agent*"]

[tool.setuptools.package-data]
"coding_agent.config" = ["config.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
